add initial bot features
This PR allows a user to trigger these commands from within a PR of a
Hubcast-enabled repo:

- `/hubcast help`: see this message
- `/hubcast approve`: sync this pull request with the destination repository and trigger a new pipeline
- `/hubcast run pipeline`: request a new run of the GitLab CI pipeline for any reason

---

These permissions are required for the Github app:

    - read: repo contents
    - read and write: pull requests (read pull request refs and write to pull request contents)
    - read and write: checks
    - read: issues (subscribe to issue/PR comments)

I'm looking for feedback on code style and modularity. I think there's a
lot of opportunity to consolidate shared logic but I'm not 100% sure of
the way you'd like to go about it.

closes #121