GitHub bot functionality
Adds features to interact with Hubcast within the context of a PR on
github.com.

The operator of a Hubcast instance sets `HC_GH_BOT_USER` to a username
of a bot to tag in a comment. For example: `@lc-hubcast help`.

The bot will respond to the following commands:
- `help`: see this message
- `approve`: sync this pull request with the destination repository and trigger a new pipeline
- `run pipeline`: request a new run of the GitLab CI pipeline for any reason

---

These permissions are required for the Github app:

    - read: repo contents
    - read and write: pull requests (read pull request refs and write to pull request contents)
    - read and write: checks
    - read: issues (subscribe to issue/PR comments)

closes #121