Add link to Gitlab pipeline URL to Github check status
Example: https://github.com/cmelone/hubcast-test/pull/1/checks?check_run_id=42994357089

Using the [check-run](https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#create-a-check-run) API, add a link to the Gitlab pipeline for the check associated with the ref. If the pipeline is canceled, retried, or failed, it will update.

While the [commit status](https://docs.github.com/en/rest/commits/statuses?apiVersion=2022-11-28#create-a-commit-status) API allows for a one-click solution within the checks table of a PR, it's not as flexible as using a check.