LDAP account map
closes #155

This PR implements an LDAP account mapper. Given the URI of an LDAP
server and the base to search from, it will filter by an input attribute
and return the value of the output attribute.

It supports GSSAPI authentication (kerberos) and simple bind
(username/password).

I've tested this on Dane with input=employeeNumber output=uid.
Eventually, the input will be githubId but this interface allows
arbitrary inputs and outputs.

I have notes documenting the specifics of each parameter for the mapper,
but I'll wait until we have a PR that fulfills #157.

TODO:
- [] test simple bind auth when we get LC service account