Allow downscoping to user identities on destination Gitlab forges
This PR implements the Gitlab impersonation token feature into Hubcast.
By configuring Hubcast with a Gitlab admin token, Hubcast will
automatically downscope to match the identity of the requesting user, so
they are marked as responsible for actions such as webhook creation and
any repo read/writes.

A couple caveats:
- Gitlab does not allow granular control of when these impersonation
  tokens expire. Currently, we set them to expire 1 day in advance.
- Because Gitlab scopes and roles interact to create the permissions
  structure for any actions, we will need to document some edge cases
  that users may run into. Any users who wish to push from a source forge to a destination forge will need to have defined read/write access to the destination repo. For Gitlab this means being at least a "developer." Additionally, only users with the "maintainer" role can create webhooks.