Support private source and destination repos
closes #158

We can detect if a source repo is private without an additional API call, in which case we'll get a
github token and pass a username and password to the `fetch_pack`
request in `sync_pr` and `sync_branch`.

We cannot detect if a destination repository is private without making
another API call. Because we're grabbing a gitlab token either way, I
just move that call up and authenticate with the credentials to the
`ls_remote` calls.

I think the overhead of checking whether a gitlab repo is private is
higher than authenticating each time, but we can look into that.